move CircleguardWindow and threading hacks out of gui.py
